microsoft-azure-functions/requirements.txt
azure-functions
requests~=2.32.3
pillow~=11.1.0